// src/app/models/cliente.ts
export class Cliente {
  SC_ID: number;
  SC_NOME: string;
  SC_PIVA?: string;
  SC_COD_FISCALE?: string;
  SC_INDIRIZZO?: string;
  SC_EMAIL?: string;
  SC_TELEFONO?: string;
  SC_REFERENTE_NOME?: string;
  SC_TEL_REFERENTE?: string;
  SC_CREATOR_ID?: number;
  SC_TS?: Date;
  SC_DELETED?: number;
}